+++ src/array.mjs
@@ -61,3 +61,4 @@ export { default as array_intersect_ukey } from './array/array_intersect_ukey.mj
 export { default as array_intersect_key } from './array/array_intersect_key.mjs';
 export { default as sort } from './array/sort.mjs';
 export { default as array_multisort } from './array/array_multisort.mjs';
+export { default as rsort } from './array/rsort.mjs';