+++ tests/array.global.test.mjs
@@ -65,4 +65,5 @@ test('array.global', () => {
     expect(array_intersect_key).not.toBeUndefined();
     expect(sort).not.toBeUndefined();
     expect(array_multisort).not.toBeUndefined();
+    expect(rsort).not.toBeUndefined();
 });